// src/app/models/User.ts
// export interface User {
//     username: string;
//     email: string;
//     password: string;
//     roles: string;
//     createdAt?: string;
//     updatedAt?: string;
//     _id?: string;
// }

export class User {
    username: string;
    email: string;
    password: string;
    rol: string;
    createdAt?: string;
    updatedAt?: string;
    _id?: string;
    imagen?: string;
}